✨ feat: Add new file `index.tsx` for settings page
This commit adds a new file `index.tsx` in the `platforms/next/pages/settings` directory. The file includes an export of the `getServerSideProps` function from `'multiplatform.one/next'` and imports the `Settings` component from `'app/screens/settings'`. Finally, it exports the `Settings` component as the default export for the file.