✨ feat: Add background image to HomeScreen
This commit adds a background image to the HomeScreen component. The new background image is `background1.png` and is imported from `app/assets/background1.png`. The image is applied as the background of the `YStack` component using inline styles. The background image is set to cover the entire screen and has a minimum height of 100vh.