♻️ refactor: add slide animations to contact screen
Added slide animations to the contact screen in order to enhance user experience. The contact image now slides in from the top, and the text on the right side slides in from the right with a delay of 300 milliseconds. The social media buttons also have slide animations, sliding in from the left with different delays. Lastly, the email button has a slide animation from the left with a delay of 400 milliseconds, and the final text at the bottom slides in from the bottom with a delay of 500 milliseconds.