✨ feat: Update SettingsInfo component with background image and animation
This commit updates the SettingsInfo component by adding a background image and animation. The backgroundImage property is set to 'app/assets/background1.png'. The component now renders a YStack with a style that includes the backgroundImage, backgroundSize, minHeight, and padding. Inside the YStack, there is a YStack with spacing of 6. The H2 component has the color set to '#fff' and displays the text "About Me". There are three Text components with fontSize of 24, fontStyle of italic, and color of '#ddd'. The first Text component contains information about an Experienced Software Engineer with proficiency in various technologies. The second Text component mentions basic knowledge of Next.Js and Nest.Js for Full-Stack Development as well as DevOps skills. The third Text component highlights experience in Custom-Debian Operating Systems, Linux environments, and relevant tools. Lastly, there is a Text component with marginTop of 6, fontSize of 28, fontStyle of italic, and color of '#fff', which describes the individual's freelancing activities and interests. At the end, a LinkStack component wraps a Button component, which allows users to get in touch. The animation is achieved using the Zoom and Fade components from the react-reveal library.